+++ library/src/main/java/com/getstream/sdk/chat/adapter/viewholder/attachment/AttachmentViewHolderMedia.kt
@@ -5,7 +5,8 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.core.content.ContextCompat
 import androidx.core.view.isVisible
-import com.bumptech.glide.Glide
+import coil.api.load
+import com.getstream.sdk.chat.ImageLoader
 import com.getstream.sdk.chat.R
 import com.getstream.sdk.chat.adapter.AttachmentListItem
 import com.getstream.sdk.chat.adapter.MessageListItem.MessageItem
@@ -66,11 +67,15 @@ class AttachmentViewHolderMedia(
         val thumbUrl = attachment.thumbUrl
         val imageUrl = if (thumbUrl.isNullOrEmpty()) attachment.imageUrl else thumbUrl
 
-        Glide.with(context)
-            .asDrawable()
-            .load(imageUrl)
-            .placeholder(R.drawable.stream_placeholder)
-            .into(binding.ivMediaThumb)
+        binding.ivMediaThumb.load(imageUrl, ImageLoader.getImageLoaderWithGifSupport(context)) {
+            placeholder(R.drawable.stream_placeholder)
+            listener(
+                onStart = { binding.progressBar.isVisible = true },
+                onSuccess = { _, _ -> binding.progressBar.isVisible = false },
+                onError = { _, _ -> binding.progressBar.isVisible = false },
+                onCancel = { binding.progressBar.isVisible = false },
+            )
+        }
 
         if (messageItem.message.type != ModelType.message_ephemeral) {
             binding.tvMediaTitle.text = attachment.title