+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # To be released:
 
+- Replace Glide with Coil in AttachmentViewHolderMedia (Fix GIFs loading issues)
+
 # Oct 14th, 2020 - 4.3.0-beta-6
 - Update to Kotlin 1.4.10
 - Fix Typing view behavior